replace int32 to int64 in msg createGiveaway

+++ x/giveaway/types/message_create_giveaway.go
@@ -9,7 +9,7 @@ const TypeMsgCreateGiveaway = "create_giveaway"
 
 var _ sdk.Msg = &MsgCreateGiveaway{}
 
-func NewMsgCreateGiveaway(creator string, duration int32, name string, prize *Prize) *MsgCreateGiveaway {
+func NewMsgCreateGiveaway(creator string, duration int64, name string, prize *Prize) *MsgCreateGiveaway {
 	return &MsgCreateGiveaway{
 		Creator:  creator,
 		Duration: duration,
